Refactors utils module
